Extract handler set from MouseRegion
